chore: move query to new file that kristin set up and delete old one
